chore: add GitHub action for stale issues and PRs (#174)
This PR enables the `notation-core-go` repo to run stale action at 1:30
every day to label or close stale PRs and issues. See guideline
https://github.com/marketplace/actions/close-stale-issues

This is the definition for stale PRs or issues that we discussed during
community call and to be updated in [contributing
PR](https://github.com/notaryproject/.github/pull/25).

"A stale issue is one that remains inactive or without updates for a
period of 60 days. A stale pull request (PR) is one that remains
inactive or without updates for a period of 45 days. When an issue or PR
becomes stale, it is labelled as `stale`. Normally maintainers will
comment on stale issues or PRs to prompt participants to take action. If
there is no activity for additional 30 days, this issue or PR will be
closed. If an update/comment occur on stale issues or pull requests, the
stale label will be removed, and the timer will restart"

Signed-off-by: Yi Zha <yizha1@microsoft.com>

---------

Signed-off-by: Yi Zha <yizha1@microsoft.com>